+++ src/main/java/frc/robot/poseestimator/IPoseEstimator.java
@@ -1,8 +1,8 @@
-package frc.robot.poseestimation.poseestimator;
+package frc.robot.poseestimator;
 
 import edu.wpi.first.math.geometry.Pose2d;
-import frc.robot.poseestimation.observations.OdometryObservation;
-import frc.robot.poseestimation.observations.VisionObservation;
+import frc.robot.poseestimator.observations.OdometryObservation;
+import frc.robot.poseestimator.observations.VisionObservation;
 
 import java.util.List;
