+++ src/main/java/frc/robot/poseestimator/IVisionEstimator.java
@@ -1,7 +1,7 @@
-package frc.robot.poseestimation.poseestimator;
+package frc.robot.poseestimator;
 
 import edu.wpi.first.math.geometry.Pose2d;
-import frc.robot.poseestimation.observations.VisionObservation;
+import frc.robot.poseestimator.observations.VisionObservation;
 
 import java.util.List;
 import java.util.Optional;
