+++ src/main/java/frc/robot/poseestimator/observations/OdometryObservation.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimation.observations;
+package frc.robot.poseestimator.observations;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveDriveWheelPositions;
