+++ src/main/java/frc/robot/vision/limelights/LimelightFiltererConfig.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimation.limelights;
+package frc.robot.vision.limelights;
 
 
 public record LimelightFiltererConfig(String logPath, String hardwareLogPath, String[] limelightsNames) {}
