+++ src/main/java/frc/robot/vision/limelights/LimelightRawData.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimation.limelights;
+package frc.robot.vision.limelights;
 
 import edu.wpi.first.math.geometry.Pose3d;
